MDEV-21117: refine the server binlog-based recovery for semisync
Problem:
=======
When the semisync master is crashed and restarted as slave it could
recover transactions that former slaves never have seen.
A existed method existed to clear out all prepared transactions
with --tc-heuristic-recover=rollback but that does not care to adjust
binlog accordingly.

Fix:
===
The binlog-based recovery is made to concern of the semisync role of
post-crash restarted server.
No changes in behaviour is done to non-semisync servers or the semisync master.
when the restarted server is enabled as the semisync slave
prepared transactions are rolled back and binlog is truncated
to synchronize itself with the engines.
In case of a partically committed (that is in one of the engines)
transaction it gets committed.
It's guaranteed no committed including partially transactions
exist beyond the truncate position.
In case there exists a non-transactional replication event past the
computed truncate position the recovery ends with an error.